bug fix isOpen()
